Refactor model building, add e2e section for inline objects.
