--- end_to_end_tests/golden-record/my_test_api_client/api/default/ping_ping_get.py
@@ -1,90 +1,0 @@
-from typing import Any, Dict, Optional
-
-import httpx
-
-from ...client import Client
-from ...types import Response
-
-
-def _get_kwargs(
-    *,
-    client: Client,
-) -> Dict[str, Any]:
-    url = "{}/ping".format(client.base_url)
-
-    headers: Dict[str, Any] = client.get_headers()
-
-    return {
-        "url": url,
-        "headers": headers,
-        "cookies": client.get_cookies(),
-        "timeout": client.get_timeout(),
-    }
-
-
-def _parse_response(*, response: httpx.Response) -> Optional[bool]:
-    if response.status_code == 200:
-        return bool(response.text)
-    return None
-
-
-def _build_response(*, response: httpx.Response) -> Response[bool]:
-    return Response(
-        status_code=response.status_code,
-        content=response.content,
-        headers=response.headers,
-        parsed=_parse_response(response=response),
-    )
-
-
-def sync_detailed(
-    *,
-    client: Client,
-) -> Response[bool]:
-    kwargs = _get_kwargs(
-        client=client,
-    )
-
-    response = httpx.get(
-        **kwargs,
-    )
-
-    return _build_response(response=response)
-
-
-def sync(
-    *,
-    client: Client,
-) -> Optional[bool]:
-    """ A quick check to see if the system is running  """
-
-    return sync_detailed(
-        client=client,
-    ).parsed
-
-
-async def asyncio_detailed(
-    *,
-    client: Client,
-) -> Response[bool]:
-    kwargs = _get_kwargs(
-        client=client,
-    )
-
-    async with httpx.AsyncClient() as _client:
-        response = await _client.get(**kwargs)
-
-    return _build_response(response=response)
-
-
-async def asyncio(
-    *,
-    client: Client,
-) -> Optional[bool]:
-    """ A quick check to see if the system is running  """
-
-    return (
-        await asyncio_detailed(
-            client=client,
-        )
-    ).parsed
