+++ end_to_end_tests/golden-record/my_test_api_client/api/tests/test_inline_objects.py
@@ -1,0 +1,68 @@
+from typing import Any, Dict
+
+import httpx
+
+from ...client import Client
+from ...types import Response
+
+
+def _get_kwargs(
+    *,
+    client: Client,
+    json_body: Dict[Any, Any],
+) -> Dict[str, Any]:
+    url = "{}/tests/inline_objects".format(client.base_url)
+
+    headers: Dict[str, Any] = client.get_headers()
+
+    json_json_body = json_body
+
+    return {
+        "url": url,
+        "headers": headers,
+        "cookies": client.get_cookies(),
+        "timeout": client.get_timeout(),
+        "json": json_json_body,
+    }
+
+
+def _build_response(*, response: httpx.Response) -> Response[None]:
+    return Response(
+        status_code=response.status_code,
+        content=response.content,
+        headers=response.headers,
+        parsed=None,
+    )
+
+
+def sync_detailed(
+    *,
+    client: Client,
+    json_body: Dict[Any, Any],
+) -> Response[None]:
+    kwargs = _get_kwargs(
+        client=client,
+        json_body=json_body,
+    )
+
+    response = httpx.post(
+        **kwargs,
+    )
+
+    return _build_response(response=response)
+
+
+async def asyncio_detailed(
+    *,
+    client: Client,
+    json_body: Dict[Any, Any],
+) -> Response[None]:
+    kwargs = _get_kwargs(
+        client=client,
+        json_body=json_body,
+    )
+
+    async with httpx.AsyncClient() as _client:
+        response = await _client.post(**kwargs)
+
+    return _build_response(response=response)
