+++ openapi_python_client/parser/openapi.py
@@ -8,6 +8,7 @@ from pydantic import ValidationError
 from .. import schema as oai
 from .. import utils
 from .errors import GeneratorError, ParseError, PropertyError
+from .model import Model, model_from_data
 from .properties import EnumProperty, Property, property_from_data
 from .reference import Reference
 from .responses import ListRefResponse, RefResponse, Response, response_from_data
@@ -221,57 +222,6 @@ class Endpoint:
 
 
 @dataclass
-class Model:
-    """
-    A data model used by the API- usually a Schema with type "object".
-
-    These will all be converted to dataclasses in the client
-    """
-
-    reference: Reference
-    required_properties: List[Property]
-    optional_properties: List[Property]
-    description: str
-    relative_imports: Set[str]
-
-    @staticmethod
-    def from_data(*, data: oai.Schema, name: str) -> Union["Model", ParseError]:
-        """A single Model from its OAI data
-
-        Args:
-            data: Data of a single Schema
-            name: Name by which the schema is referenced, such as a model name.
-                Used to infer the type name if a `title` property is not available.
-        """
-        required_set = set(data.required or [])
-        required_properties: List[Property] = []
-        optional_properties: List[Property] = []
-        relative_imports: Set[str] = set()
-
-        ref = Reference.from_ref(data.title or name)
-
-        for key, value in (data.properties or {}).items():
-            required = key in required_set
-            p = property_from_data(name=key, required=required, data=value)
-            if isinstance(p, ParseError):
-                return p
-            if required:
-                required_properties.append(p)
-            else:
-                optional_properties.append(p)
-            relative_imports.update(p.get_imports(prefix=".."))
-
-        model = Model(
-            reference=ref,
-            required_properties=required_properties,
-            optional_properties=optional_properties,
-            relative_imports=relative_imports,
-            description=data.description or "",
-        )
-        return model
-
-
-@dataclass
 class Schemas:
     """ Contains all the Schemas (references) for an OpenAPI document """
 
@@ -296,7 +246,7 @@ class Schemas:
                     nullable=data.nullable,
                 )
                 continue
-            s = Model.from_data(data=data, name=name)
+            s = model_from_data(data=data, name=name)
             if isinstance(s, ParseError):
                 result.errors.append(s)
             else:
@@ -311,7 +261,8 @@ class GeneratorData:
     title: str
     description: Optional[str]
     version: str
-    schemas: Schemas
+    models: Dict[str, Model]
+    errors: List[ParseError]
     endpoint_collections_by_tag: Dict[str, EndpointCollection]
     enums: Dict[str, EnumProperty]
 
@@ -334,6 +285,7 @@ class GeneratorData:
             description=openapi.info.description,
             version=openapi.info.version,
             endpoint_collections_by_tag=endpoint_collections_by_tag,
-            schemas=schemas,
+            models=schemas.models,
+            errors=schemas.errors,
             enums=enums,
         )
