+++ tests/test_parser/test_openapi.py
@@ -1,8 +1,5 @@
-from pydantic import ValidationError
-from pydantic.error_wrappers import ErrorWrapper
-
 import openapi_python_client.schema as oai
-from openapi_python_client import GeneratorError, utils
+from openapi_python_client import GeneratorError
 from openapi_python_client.parser.errors import ParseError
 
 MODULE_NAME = "openapi_python_client.parser.openapi"
@@ -31,7 +28,8 @@ class TestGeneratorData:
             description=openapi.info.description,
             version=openapi.info.version,
             endpoint_collections_by_tag=EndpointCollection.from_data.return_value,
-            schemas=Schemas.build.return_value,
+            errors=Schemas.build.return_value.errors,
+            models=Schemas.build.return_value.models,
             enums=get_all_enums.return_value,
         )
 
@@ -42,7 +40,8 @@ class TestGeneratorData:
         generator_data = GeneratorData.from_dict(in_dict)
 
         Schemas.build.assert_not_called()
-        assert generator_data.schemas == Schemas()
+        assert generator_data.models == Schemas.return_value.models
+        assert generator_data.errors == Schemas.return_value.errors
 
     def test_from_dict_invalid_schema(self, mocker):
         Schemas = mocker.patch(f"{MODULE_NAME}.Schemas")
@@ -67,87 +66,9 @@ class TestGeneratorData:
         Schemas.assert_not_called()
 
 
-class TestModel:
-    def test_from_data(self, mocker):
-        from openapi_python_client.parser.properties import Property
-
-        in_data = oai.Schema.construct(
-            title=mocker.MagicMock(),
-            description=mocker.MagicMock(),
-            required=["RequiredEnum"],
-            properties={
-                "RequiredEnum": mocker.MagicMock(),
-                "OptionalDateTime": mocker.MagicMock(),
-            },
-        )
-        required_property = mocker.MagicMock(autospec=Property)
-        required_imports = mocker.MagicMock()
-        required_property.get_imports.return_value = {required_imports}
-        optional_property = mocker.MagicMock(autospec=Property)
-        optional_imports = mocker.MagicMock()
-        optional_property.get_imports.return_value = {optional_imports}
-        property_from_data = mocker.patch(
-            f"{MODULE_NAME}.property_from_data",
-            side_effect=[required_property, optional_property],
-        )
-        from_ref = mocker.patch(f"{MODULE_NAME}.Reference.from_ref")
-
-        from openapi_python_client.parser.openapi import Model
-
-        result = Model.from_data(data=in_data, name=mocker.MagicMock())
-
-        from_ref.assert_called_once_with(in_data.title)
-        property_from_data.assert_has_calls(
-            [
-                mocker.call(name="RequiredEnum", required=True, data=in_data.properties["RequiredEnum"]),
-                mocker.call(name="OptionalDateTime", required=False, data=in_data.properties["OptionalDateTime"]),
-            ]
-        )
-        required_property.get_imports.assert_called_once_with(prefix="..")
-        optional_property.get_imports.assert_called_once_with(prefix="..")
-        assert result == Model(
-            reference=from_ref(),
-            required_properties=[required_property],
-            optional_properties=[optional_property],
-            relative_imports={
-                required_imports,
-                optional_imports,
-            },
-            description=in_data.description,
-        )
-
-    def test_from_data_property_parse_error(self, mocker):
-        in_data = oai.Schema.construct(
-            title=mocker.MagicMock(),
-            description=mocker.MagicMock(),
-            required=["RequiredEnum"],
-            properties={
-                "RequiredEnum": mocker.MagicMock(),
-                "OptionalDateTime": mocker.MagicMock(),
-            },
-        )
-        parse_error = ParseError(data=mocker.MagicMock())
-        property_from_data = mocker.patch(
-            f"{MODULE_NAME}.property_from_data",
-            return_value=parse_error,
-        )
-        from_ref = mocker.patch(f"{MODULE_NAME}.Reference.from_ref")
-
-        from openapi_python_client.parser.openapi import Model
-
-        result = Model.from_data(data=in_data, name=mocker.MagicMock())
-
-        from_ref.assert_called_once_with(in_data.title)
-        property_from_data.assert_called_once_with(
-            name="RequiredEnum", required=True, data=in_data.properties["RequiredEnum"]
-        )
-
-        assert result == parse_error
-
-
 class TestSchemas:
     def test_build(self, mocker):
-        from_data = mocker.patch(f"{MODULE_NAME}.Model.from_data")
+        from_data = mocker.patch(f"{MODULE_NAME}.model_from_data")
         in_data = {"1": mocker.MagicMock(enum=None), "2": mocker.MagicMock(enum=None), "3": mocker.MagicMock(enum=None)}
         schema_1 = mocker.MagicMock()
         schema_2 = mocker.MagicMock()
@@ -178,7 +99,7 @@ class TestSchemas:
     def test_build_enums(self, mocker):
         from openapi_python_client.parser.openapi import Schemas
 
-        from_data = mocker.patch(f"{MODULE_NAME}.Model.from_data")
+        from_data = mocker.patch(f"{MODULE_NAME}.model_from_data")
         enum_property = mocker.patch(f"{MODULE_NAME}.EnumProperty")
         in_data = {"1": mocker.MagicMock(enum=["val1", "val2", "val3"])}
